adds tls endpoint and integrates with ARPS for authZ
This commit uses EbsAuthHelpers to integrate with ARPS for authorization. It
exposes a new TLS-based endpoint using the amazon hostCert jks when running in
a secure domain. For insecure domains, such as a desktop, it uses a self-signed
cert.

Secure domains expose the TLS endpoint on 0.0.0.0 and map the insecure endpoint to
the loopback device on 127.0.0.1 so it cannot be called from outside the host.

This commit adds RIP data to the service runtime so it's able to look up the region
the service is running in for the ARPS integration.

SIM: https://sim.amazon.com/issues/Glycerin-302

cr: https://code.amazon.com/reviews/CR-63959942