Adds CapacityProvider interface to load data into memory
This change adds a CapacityProvider interface which is responsible for
loading data into Blacksmith's memory. The change contains an implementation
which loads data from a static file on disk. It also contains a capacity cache
which provides an abstraction for Blackmith over the metal servers which are
in the fleet to be used for placement.

cr: https://code.amazon.com/reviews/CR-65120503